requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0
transformers==4.33.2
torch==2.0.1
accelerate==0.23.0
gunicorn==21.2.0
httpx==0.24.1
python-multipart==0.0.6